allow ex-aod members to be accessed

+++ resources/views/member/edit-user.blade.php
@@ -54,9 +54,11 @@
         </div>
 
         @can('delete', $member)
-            {!! Form::model($member, ['method' => 'delete', 'route' => ['deleteMember', $member->clan_id]]) !!}
-            @include('member.forms.remove-member-form')
-            {!! Form::close() !!}
+            @if ($member->division)
+                {!! Form::model($member, ['method' => 'delete', 'route' => ['deleteMember', $member->clan_id]]) !!}
+                @include('member.forms.remove-member-form')
+                {!! Form::close() !!}
+            @endif
         @endcan
     </div>
